1.) Changed the fields in the countries class in trial_prediction_2 from Decimal to Float 2.) Created the trial_prediction_2/apps.py 3.) Created the trial_prediction_2 app entry in settings.py 4.) Migrated the database

+++ trial_prediction_2/models.py
@@ -3,9 +3,9 @@ from django.db import models
 # Create your models here.
 class countries(models.Model):
     country_destination = models.CharField(max_length=2)
-    lat_destination = models.DecimalField()
-    lng_destination = models.DecimalField()
-    distance_km = models.DecimalField()
+    lat_destination = models.FloatField()
+    lng_destination = models.FloatField()
+    distance_km = models.FloatField()
     destination_km2 = models.IntegerField()
     destination_language = models.CharField(max_length=3)
     language_levenshtein_distance = models.FloatField()
